+++ packages/components/src/form/ImageProfilePicker/ImageProfilePicker.stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import mdx from './ImageProfilePicker.mdx';
-import ImageProfilePicker from './ImageProfilePicker';
+import ImageProfilePicker, { IMAGE_PROFILE_PICKER_VARIANTS } from './ImageProfilePicker';
 
 export default {
   title: 'Organisms/Form/ImageProfilePicker',
@@ -16,6 +16,13 @@ export default {
   },
   argTypes: {
     onChange: { action: 'Value changed' },
+    variant: {
+      options: IMAGE_PROFILE_PICKER_VARIANTS,
+      control: { type: 'select' },
+    },
+    loading: {
+      type: 'boolean',
+    },
   },
 };
 
@@ -36,4 +43,5 @@ export const Playground = Template.bind({});
 
 Playground.args = {
   fullName: 'John Doe',
+  loading: false,
 };