+++ packages/components/src/layout/TotalLayout/TotalLayoutFooterContainer/TotalLayoutFooterContainer.js
@@ -17,6 +17,7 @@ const TotalLayoutFooterContainer = ({
   width,
   clean,
   rectRef,
+  skipOffset,
 }) => {
   const [showFooterBorder, setShowFooterBorder] = React.useState(false);
   const [footerStyles, setFooterStyles] = React.useState({});
@@ -27,6 +28,7 @@ const TotalLayoutFooterContainer = ({
       fixed,
       fullWidth,
       width,
+      skipOffset,
     },
     { name: 'TotalLayoutFooterContainer' },
   );
@@ -63,18 +65,19 @@ const TotalLayoutFooterContainer = ({
   React.useEffect(() => {
     let animationFrameId;
 
-    const updateFooterStyles = () => {
-      const clientRect = rectRef?.current?.getBoundingClientRect();
-      // Aquí actualizamos el estado o hacemos lo que sea necesario con clientRect
-      // Por ejemplo, si necesitas actualizar el estado basado en clientRect, puedes hacerlo aquí.
-      // Asegúrate de tener un estado para almacenar los estilos que quieres actualizar.
-      setFooterStyles({ left: clientRect?.left, width: clientRect?.width });
+    if (rectRef?.current) {
+      const updateFooterStyles = () => {
+        const clientRect = rectRef?.current?.getBoundingClientRect();
+        // Aquí actualizamos el estado o hacemos lo que sea necesario con clientRect
+        // Por ejemplo, si necesitas actualizar el estado basado en clientRect, puedes hacerlo aquí.
+        // Asegúrate de tener un estado para almacenar los estilos que quieres actualizar.
+        setFooterStyles({ left: clientRect?.left, width: clientRect?.width });
+
+        animationFrameId = requestAnimationFrame(updateFooterStyles);
+      };
 
       animationFrameId = requestAnimationFrame(updateFooterStyles);
-    };
-
-    animationFrameId = requestAnimationFrame(updateFooterStyles);
-
+    }
     return () => {
       cancelAnimationFrame(animationFrameId);
     };
@@ -111,6 +114,7 @@ TotalLayoutFooterContainer.propTypes = {
   width: PropTypes.number,
   clean: PropTypes.bool,
   rectRef: PropTypes.object,
+  skipOffset: PropTypes.bool,
 };
 
 export { TotalLayoutFooterContainer };