+++ packages/components/src/overlay/Drawer/Drawer.stories.js
@@ -22,10 +22,13 @@ export default {
       options: DRAWERS_SIZES,
       control: { type: 'select' },
     },
+    contentLoading: {
+      control: { type: 'boolean' },
+    },
   },
 };
 
-const Template = ({ ...props }) => {
+const Template = ({ contentLoading, ...props }) => {
   const [opened, setOpened] = React.useState(props.opened);
 
   React.useEffect(() => {
@@ -38,8 +41,12 @@ const Template = ({ ...props }) => {
         <Button onClick={() => setOpened(true)}>Open Drawer</Button>
       </Stack>
       <Drawer {...props} opened={opened} onClose={() => setOpened(false)}>
-        <Drawer.Header title="Drawer Title" />
-        <Drawer.Content>
+        <Drawer.Header title="Drawer Title">
+          <Drawer.Header.RightActions>
+            <span>Holaa</span>
+          </Drawer.Header.RightActions>
+        </Drawer.Header>
+        <Drawer.Content loading={contentLoading}>
           <ContextContainer title="Drawer Content - Part I">
             <Paragraph>
               Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
@@ -85,46 +92,6 @@ const Template = ({ ...props }) => {
               Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt
               mollit anim id est laborum.
             </Paragraph>
-            <Paragraph>
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-              incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
-              exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure
-              dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
-              Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt
-              mollit anim id est laborum.
-            </Paragraph>
-            <Paragraph>
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-              incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
-              exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure
-              dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
-              Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt
-              mollit anim id est laborum.
-            </Paragraph>
-            <Paragraph>
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-              incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
-              exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure
-              dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
-              Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt
-              mollit anim id est laborum.
-            </Paragraph>
-            <Paragraph>
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-              incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
-              exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure
-              dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
-              Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt
-              mollit anim id est laborum.
-            </Paragraph>
-            <Paragraph>
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-              incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
-              exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure
-              dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
-              Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt
-              mollit anim id est laborum.
-            </Paragraph>
           </ContextContainer>
         </Drawer.Content>
         <Drawer.Footer>
@@ -146,4 +113,5 @@ Playground.args = {
   ...DRAWER_DEFAULT_PROPS,
   opened: true,
   modalAriaLabel: 'Example of a Drawer',
+  contentLoading: false,
 };