+++ packages/leemons/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @bubbles-ui/leemons
 
+## 1.2.208
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.208
+  - @bubbles-ui/calendars@1.2.208
+  - @bubbles-ui/editors@1.2.208
+  - @bubbles-ui/icons@1.2.208
+
+## 1.2.207
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.207
+  - @bubbles-ui/calendars@1.2.207
+  - @bubbles-ui/editors@1.2.207
+  - @bubbles-ui/icons@1.2.207
+
+## 1.2.206
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.206
+  - @bubbles-ui/calendars@1.2.206
+  - @bubbles-ui/editors@1.2.206
+  - @bubbles-ui/icons@1.2.206
+
+## 1.2.205
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.205
+  - @bubbles-ui/calendars@1.2.205
+  - @bubbles-ui/editors@1.2.205
+  - @bubbles-ui/icons@1.2.205
+
+## 1.2.204
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.204
+  - @bubbles-ui/calendars@1.2.204
+  - @bubbles-ui/editors@1.2.204
+  - @bubbles-ui/icons@1.2.204
+
 ## 1.2.203
 
 ### Patch Changes